Provide a mechanism to include {bundle}/Java in the class path.
This is used by tools that put other objects and code-resources under the Java/ tree, and then reference them at runtime, expecting the classpath to include them.  My particular case was using JRuby and a bunch of custom jars and gems to go with it.

It's an edge case, but a useful one.

Usage:
    <bundleapp ... includejavaroot="true">...</bundleapp>